+++ src/core/__init__.py
@@ -2,8 +2,8 @@ import sys
 from tornado.ioloop import IOLoop
 from tornado.web import Application
 from typing import Union
-from .handler.http.http import HttpHandler, HttpSystemHandler, HttpNetworkHandler
-from .handler.websocket.websocket import WebsocketHandler
+from .handler.http.http_handler import HttpHandler, HttpSystemHandler, HttpNetworkHandler
+from .handler.websocket.websocket_handler import WebsocketHandler
 from .thread_pool import executor
