+++ src/core/handler/http/get_network_data.py
