+++ src/core/handler/websocket/get_network_data.py
