+++ src/core/handler/websocket/get_system_data.py
